GoalBased/requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-google-genai
langchain-community
PyMuPDF
google-generativeai>=0.8.0
